fix(extras): align nic_speed_gbps custom field type with netbox-proxbox plugin
The netbox-proxbox v0.0.15 plugin's _v0_0_15_release_data migration creates
the nic_speed_gbps custom field as type 'integer' at NetBox startup. The
proxbox-api bootstrap then tried to PATCH it to 'decimal', and NetBox 4.x
rejects custom-field type changes with:

    {"type": ["Changing the type of custom fields is not supported."]}

This produced HTTP 400 from /extras/extras/custom-fields/create in every
E2E run after the plugin was preloaded, blocking PR #87 CI.

Switch the bootstrap declaration to 'integer' to match the plugin migration.
The value-writing path in services/hardware_discovery.py is type-agnostic
and integer covers all real Gbps NIC link speeds (1, 10, 25, 40, 100, 400)
without precision loss.

+++ proxbox_api/routes/extras/__init__.py
@@ -697,7 +697,7 @@ async def create_custom_fields(  # noqa: C901
             },
             {
                 "object_types": ["dcim.interface"],
-                "type": "decimal",
+                "type": "integer",
                 "name": "nic_speed_gbps",
                 "label": "NIC Speed (Gbps)",
                 "description": "Negotiated link speed reported by ethtool, in Gbps",
